Extract skill source root discovery
Move skill source root enumeration and SKILL.md source discovery out of SkillInstallLayout while keeping the existing SkillInstallLayout entry points as delegating wrappers.
