+++ Packages/src/Editor/Infrastructure/SkillSetup/SkillSourceRootEnumerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6d03979eff3884a1c946386f990cf080
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
